Remove auspice dead code; consolidate util/ into utils/ (#286)
Addresses all three sub-tasks of #286.

Sub-task 1 — delete src/util/globals.js (857 lines of Nextstrain/auspice
vestige). Its only two live exports are relocated:
- controlsWidth -> src/globalStyles.js (its sole consumer)
- dataBaseURL   -> src/actions/clientDataLoader.js (its sole consumer)
(charonAPIAddress was already dead — its consumer loadData.js was deleted in
#293.) Dropped the now-obsolete jest.mock of util/globals in the
clientDataLoader test.

Sub-task 2 — resolve the src/util/ vs src/utils/ split. Four src/util/
modules had no live callers (only their own tests): computeResponsive.js,
stringHelpers.js, sets.js, perf.js — deleted with their tests. The one live
module, polyfills.js (side-effect import in src/index.js), moved to
src/utils/polyfills.js. src/util/ is removed.

Also removed the dead perf-timing machinery that only existed to instrument
the deleted perf.js: the build:client:perf / build:start:perf /
start:local:perf scripts, the babel-plugin-strip-function-call devDependency,
and the production/devtiming Babel env entries (which stripped timerStart/
timerEnd calls that no longer exist). Lockfile regenerated.

Sub-task 3 — reword CLAUDE.md "Architecture Principles": replace the
aspirational "modern patterns" bullet with an honest "pragmatic stability"
stance (class components + @connect throughout, 0 hooks; a hooks migration
would be a separate epic).

Docs (CLAUDE.md / DEVELOPMENT.md structure trees) and CHANGELOG updated.

Verified: npm run lint (0 errors), npm test (617/617), npm run build,
npm run format:check, npm run test:e2e (2/2) all pass.
