+++ DEVELOPMENT.md
@@ -206,9 +206,6 @@ npm run build
 
 # Build and immediately serve
 npm run build:start
-
-# Build with performance timing
-npm run build:start:perf
 ```
 
 ### Electron (Desktop App)
@@ -431,7 +428,6 @@ olmsted/
 │   │   ├── olmstedDB.js      # Dexie database schema
 │   │   ├── clientDataStore.js # Storage abstraction + caching
 │   │   └── fileProcessor.js  # File parsing
-│   ├── util/                 # Core application utilities
 │   ├── css/                  # Stylesheets
 │   ├── images/               # Static images
 │   └── server/               # Express server code
