+++ package.json
@@ -15,15 +15,12 @@
     "clean": "rimraf _deploy/dist",
     "server": "node server.dist.js",
     "build:client": "NODE_ENV=production BABEL_ENV=production webpack --config webpack.config.prod.js",
-    "build:client:perf": "NODE_ENV=production BABEL_ENV=productiontiming webpack --config webpack.config.prod.js",
     "build:electron": "NODE_ENV=production BABEL_ENV=production webpack --config webpack.config.electron.js",
     "build:server": "NODE_ENV=production BABEL_ENV=production webpack --config webpack.config.server.js",
     "build": "npm run clean && npm run build:client && npm run build:server",
     "build:start": "npm run build && npm run server",
-    "build:start:perf": "npm run clean && npm run build:client:perf && npm run build:server && npm run server",
     "start": "BABEL_ENV=dev ./node_modules/.bin/babel-node server.js dev",
     "start:local": "BABEL_ENV=dev ./node_modules/.bin/babel-node server.js dev localData",
-    "start:local:perf": "BABEL_ENV=devtiming ./node_modules/.bin/babel-node server.js dev localData",
     "lint": "eslint src",
     "format": "prettier --write \"src/**/*.{js,jsx,json,css,md}\"",
     "format:check": "prettier --check \"src/**/*.{js,jsx,json,css,md}\"",
@@ -104,7 +101,6 @@
     "@testing-library/react": "^16.3.2",
     "babel-jest": "^30.3.0",
     "babel-loader": "^10.1.1",
-    "babel-plugin-strip-function-call": "^1.0.2",
     "chokidar": "^5.0.0",
     "compression-webpack-plugin": "^12.0.0",
     "css-loader": "^7.1.4",
