--- src/util/__tests__/stringHelpers.test.js
@@ -1,100 +1,0 @@
-import { prettyString, formatURLString } from "../stringHelpers";
-
-describe("prettyString", () => {
-  describe("falsy inputs", () => {
-    it("returns empty string for undefined", () => {
-      expect(prettyString(undefined)).toBe("");
-    });
-
-    it("returns empty string for null", () => {
-      expect(prettyString(null)).toBe("");
-    });
-
-    it("returns empty string for empty string", () => {
-      expect(prettyString("")).toBe("");
-    });
-
-    it("handles 0 as a number (not falsy)", () => {
-      expect(prettyString(0)).not.toBe("");
-    });
-  });
-
-  describe("string inputs", () => {
-    it("replaces underscores with spaces and camelCases", () => {
-      expect(prettyString("hello_world")).toBe("Hello World");
-    });
-
-    it("uppercases country abbreviations (usvi, usa, uk)", () => {
-      expect(prettyString("usa")).toBe("USA");
-      expect(prettyString("usvi")).toBe("USVI");
-      expect(prettyString("uk")).toBe("UK");
-    });
-
-    it("trims long strings when trim option is set", () => {
-      expect(prettyString("a_very_long_string", { trim: 5 })).toBe("A Ver...");
-    });
-
-    it("does not trim when string is shorter than trim limit", () => {
-      expect(prettyString("short", { trim: 10 })).toBe("Short");
-    });
-
-    it("disables camelCase when option is false", () => {
-      expect(prettyString("hello_world", { camelCase: false })).toBe("hello world");
-    });
-
-    it("removes commas when removeComma is true", () => {
-      // camelCase runs first (joins comma-adjacent chars as words), then commas are removed
-      expect(prettyString("one,two,three", { removeComma: true })).toBe("Onetwothree");
-    });
-
-    it("strips 'et al' variations when stripEtAl is true", () => {
-      expect(prettyString("smith et al.", { stripEtAl: true })).toBe("Smith ");
-    });
-  });
-
-  describe("number inputs", () => {
-    it("formats numbers with appropriate decimal places", () => {
-      const result = prettyString(123.456);
-      expect(typeof result).toBe("string");
-      expect(result).toMatch(/^\d+\.\d+$/);
-    });
-
-    it("adds multiplier symbol when multiplier option is true", () => {
-      const result = prettyString(5, { multiplier: true });
-      expect(result).toContain("\u00D7");
-    });
-
-    it("does not add multiplier symbol by default", () => {
-      const result = prettyString(5);
-      expect(result).not.toContain("\u00D7");
-    });
-  });
-
-  describe("other types", () => {
-    it("returns non-string non-number values as-is", () => {
-      const obj = { foo: "bar" };
-      expect(prettyString(obj)).toBe(obj);
-    });
-  });
-});
-
-describe("formatURLString", () => {
-  it("converts https_ prefix to https:", () => {
-    expect(formatURLString("https_//example.com")).toBe("https://example.com");
-  });
-
-  it("leaves normal URLs unchanged", () => {
-    expect(formatURLString("https://example.com")).toBe("https://example.com");
-  });
-
-  it("handles strings without URL prefixes", () => {
-    expect(formatURLString("example.com")).toBe("example.com");
-  });
-
-  it("handles http_ prefix (note: source has a bug — uses https_ regex for http case)", () => {
-    // The source code has: url.replace(/^https_/, "http:") for the http_ case
-    // which means http_ won't actually be replaced. Testing actual behavior.
-    const result = formatURLString("http_//example.com");
-    expect(result).toBe("http_//example.com");
-  });
-});
